feat: Transaction reputation
- Se agrego a las transacciones el atributo reputacion
- Se arreglo el tipo de verbos de los endpoints

+++ src/main/java/ar/edu/unq/desapp/grupoD022021/backenddesappapi/dto/ActivityDto.java
@@ -84,7 +84,7 @@ public class ActivityDto  {
         this.numberOperations = numberOperations;
     }
 
-    public double getReputation() {
+    public int getReputation() {
         return reputation;
     }
